move ai-insights and insight components to observability_agent_builder
