+++ package.json
@@ -180,7 +180,6 @@
     "@kbn/ai-assistant-cta": "link:x-pack/platform/packages/shared/ai-assistant/ai-assistant-cta",
     "@kbn/ai-assistant-icon": "link:x-pack/platform/packages/shared/ai-assistant/icon",
     "@kbn/ai-assistant-management-plugin": "link:src/platform/plugins/shared/ai_assistant_management/selection",
-    "@kbn/ai-insights": "link:x-pack/platform/packages/shared/ai-insights",
     "@kbn/ai-security-labs-content": "link:x-pack/solutions/security/packages/ai-security-labs-content",
     "@kbn/ai-tools": "link:x-pack/platform/packages/shared/kbn-ai-tools",
     "@kbn/aiops-change-point-detection": "link:x-pack/platform/packages/private/ml/aiops_change_point_detection",
