+++ tsconfig.base.json
@@ -34,8 +34,6 @@
       "@kbn/ai-assistant-icon/*": ["x-pack/platform/packages/shared/ai-assistant/icon/*"],
       "@kbn/ai-assistant-management-plugin": ["src/platform/plugins/shared/ai_assistant_management/selection"],
       "@kbn/ai-assistant-management-plugin/*": ["src/platform/plugins/shared/ai_assistant_management/selection/*"],
-      "@kbn/ai-insights": ["x-pack/platform/packages/shared/ai-insights"],
-      "@kbn/ai-insights/*": ["x-pack/platform/packages/shared/ai-insights/*"],
       "@kbn/ai-security-labs-content": ["x-pack/solutions/security/packages/ai-security-labs-content"],
       "@kbn/ai-security-labs-content/*": ["x-pack/solutions/security/packages/ai-security-labs-content/*"],
       "@kbn/ai-tools": ["x-pack/platform/packages/shared/kbn-ai-tools"],
