--- x-pack/platform/packages/shared/ai-insights/moon.yml
@@ -1,48 +1,0 @@
-# This file is generated by the @kbn/moon package. Any manual edits will be erased!
-#  To extend this, write your extensions/overrides to 'moon.extend.yml'
-#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/ai-insights'
-
-$schema: https://moonrepo.dev/schemas/project.json
-id: '@kbn/ai-insights'
-type: unknown
-owners:
-  defaultOwner: '@elastic/obs-ai-team'
-toolchain:
-  default: node
-language: typescript
-project:
-  name: '@kbn/ai-insights'
-  description: Moon project for @kbn/ai-insights
-  channel: ''
-  owner: '@elastic/obs-ai-team'
-  metadata:
-    sourceRoot: x-pack/platform/packages/shared/ai-insights
-dependsOn:
-  - '@kbn/i18n'
-  - '@kbn/licensing-types'
-  - '@kbn/kibana-react-plugin'
-tags:
-  - shared-browser
-  - package
-  - prod
-  - group-platform
-  - shared
-  - jest-unit-tests
-fileGroups:
-  src:
-    - '**/*.ts'
-    - '**/*.tsx'
-    - '!target/**/*'
-tasks:
-  jest:
-    args:
-      - '--config'
-      - $projectRoot/jest.config.js
-    inputs:
-      - '@group(src)'
-  jestCI:
-    args:
-      - '--config'
-      - $projectRoot/jest.config.js
-    inputs:
-      - '@group(src)'
