--- x-pack/platform/packages/shared/ai-insights/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "@kbn/ai-insights",
-  "description": "AI Insights UI components",
-  "private": true,
-  "version": "1.0.0",
-  "license": "Elastic License 2.0"
-}
