--- x-pack/platform/packages/shared/ai-insights/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@kbn/tsconfig-base/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "target/types",
-    "types": ["jest", "node", "react"]
-  },
-  "include": ["**/*.ts", "**/*.tsx"],
-  "exclude": ["target/**/*"],
-  "kbn_references": [
-    "@kbn/i18n",
-    "@kbn/licensing-types",
-    "@kbn/kibana-react-plugin",
-  ]
-}
