+++ x-pack/solutions/observability/plugins/apm/moon.yml
@@ -152,7 +152,6 @@ dependsOn:
   - '@kbn/core-chrome-layout-constants'
   - '@kbn/zod'
   - '@kbn/onechat-plugin'
-  - '@kbn/ai-insights'
   - '@kbn/observability-agent-builder-plugin'
   - '@kbn/server-route-repository-client'
   - '@kbn/ai-assistant-common'
