+++ x-pack/solutions/observability/plugins/apm/public/components/app/error_group_details/error_sampler/error_sample_detail.tsx
@@ -50,7 +50,7 @@ import { ErrorTabKey, getTabs } from './error_tabs';
 import { ErrorUiActionsContextMenu } from './error_ui_actions_context_menu';
 import { SampleSummary } from './sample_summary';
 import { ErrorSampleContextualInsight } from './error_sample_contextual_insight';
-import { ErrorSampleAiInsight } from './error_sample_ai_insight';
+import { useTimeRange } from '../../../../hooks/use_time_range';
 import { getComparisonEnabled } from '../../../shared/time_comparison/get_comparison_enabled';
 import { buildUrl } from '../../../../utils/build_url';
 import { OpenErrorInDiscoverButton } from '../../../shared/links/discover_links/open_error_in_discover_button';
@@ -89,7 +89,9 @@ export function ErrorSampleDetails({
     urlParams: { detailTab, offset, comparisonEnabled },
   } = useLegacyUrlParams();
 
-  const { uiActions, core } = useApmPluginContext();
+  const { uiActions, core, observabilityAgentBuilder } = useApmPluginContext();
+
+  const ErrorSampleAiInsight = observabilityAgentBuilder?.getErrorSampleAIInsight();
 
   const router = useApmRouter();
 
@@ -99,7 +101,8 @@ export function ErrorSampleDetails({
     '/mobile-services/{serviceName}/errors-and-crashes/crashes/{groupId}'
   );
 
-  const { kuery } = query;
+  const { kuery, rangeFrom, rangeTo, environment } = query;
+  const { start, end } = useTimeRange({ rangeFrom, rangeTo });
 
   const loadingErrorSamplesData = isPending(errorSamplesFetchStatus);
   const loadingErrorData = isPending(errorFetchStatus);
@@ -168,7 +171,7 @@ export function ErrorSampleDetails({
   const method = errorOrTransactionHttp?.http?.request?.method;
   const status = errorOrTransactionHttp?.http?.response?.status_code;
   const userAgent = errorOrTransactionUserAgent;
-  const environment = error.service.environment;
+  const errorEnvironment = error.service.environment;
   const serviceVersion = error.service.version;
   const isUnhandled = error.error.exception?.[0]?.handled === false;
 
@@ -249,14 +252,14 @@ export function ErrorSampleDetails({
                 </TransactionDetailLink>
               </EuiToolTip>
             ),
-            environment ? (
+            errorEnvironment ? (
               <EuiToolTip
                 content={i18n.translate('xpack.apm.errorSampleDetails.serviceEnvironment', {
                   defaultMessage: 'Environment',
                 })}
               >
                 <EuiBadge color="hollow" tabIndex={0}>
-                  {environment}
+                  {errorEnvironment}
                 </EuiBadge>
               </EuiToolTip>
             ) : null,
@@ -292,7 +295,15 @@ export function ErrorSampleDetails({
         <SampleSummary error={error} />
       )}
 
-      <ErrorSampleAiInsight error={error} />
+      {ErrorSampleAiInsight && error && (
+        <ErrorSampleAiInsight
+          errorId={error.error.id}
+          serviceName={error.service.name}
+          start={start}
+          end={end}
+          environment={environment}
+        />
+      )}
       <ErrorSampleContextualInsight error={error} transaction={transaction} />
 
       <EuiTabs>
