+++ x-pack/solutions/observability/plugins/apm/public/plugin.ts
@@ -78,6 +78,7 @@ import type { FieldsMetadataPublicStart } from '@kbn/fields-metadata-plugin/publ
 import type { SharePublicStart } from '@kbn/share-plugin/public/plugin';
 import type { ApmSourceAccessPluginStart } from '@kbn/apm-sources-access-plugin/public';
 import type { OnechatPluginStart } from '@kbn/onechat-plugin/public';
+import type { ObservabilityAgentBuilderPluginPublicStart } from '@kbn/observability-agent-builder-plugin/public';
 import type { CasesPublicStart } from '@kbn/cases-plugin/public';
 import type {
   DiscoverSharedPublicSetup,
@@ -169,6 +170,7 @@ export interface ApmPluginStartDeps {
   notifications: NotificationsStart;
   discoverShared: DiscoverSharedPublicStart;
   onechat?: OnechatPluginStart;
+  observabilityAgentBuilder?: ObservabilityAgentBuilderPluginPublicStart;
 }
 
 const applicationsTitle = i18n.translate('xpack.apm.navigation.rootTitle', {
