+++ x-pack/solutions/observability/plugins/apm/tsconfig.json
@@ -150,7 +150,6 @@
     "@kbn/core-chrome-layout-constants",
     "@kbn/zod",
     "@kbn/onechat-plugin",
-    "@kbn/ai-insights",
     "@kbn/observability-agent-builder-plugin",
     "@kbn/server-route-repository-client",
     "@kbn/ai-assistant-common",
