+++ x-pack/solutions/observability/plugins/observability/moon.yml
@@ -148,7 +148,6 @@ dependsOn:
   - '@kbn/ai-assistant-common'
   - '@kbn/management-settings-ids'
   - '@kbn/onechat-plugin'
-  - '@kbn/ai-insights'
   - '@kbn/observability-agent-builder-plugin'
   - '@kbn/server-route-repository-client'
 tags:
