+++ x-pack/solutions/observability/plugins/observability/public/navigation_tree.ts
@@ -31,7 +31,7 @@ const LazyIconProductCloudInfra = lazy(() =>
   }))
 );
 const LazyAgentBuilderIcon = lazy(() =>
-  import('@kbn/ai-insights/src/icons/robot_icon').then(({ RobotIcon }) => ({
+  import('@kbn/observability-agent-builder-plugin/public').then(({ RobotIcon }) => ({
     default: RobotIcon,
   }))
 );
