+++ x-pack/solutions/observability/plugins/observability/tsconfig.json
@@ -143,7 +143,6 @@
     "@kbn/ai-assistant-common",
     "@kbn/management-settings-ids",
     "@kbn/onechat-plugin",
-    "@kbn/ai-insights",
     "@kbn/observability-agent-builder-plugin",
     "@kbn/server-route-repository-client",
   ],
