+++ x-pack/solutions/observability/plugins/observability_agent_builder/moon.yml
@@ -48,7 +48,6 @@ dependsOn:
   - '@kbn/management-settings-ids'
   - '@kbn/logging'
   - '@kbn/ai-assistant-common'
-  - '@kbn/ai-insights'
   - '@kbn/kibana-react-plugin'
   - '@kbn/shared-ux-utility'
   - '@kbn/discover-shared-plugin'
