+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/components/ai_insight/ai_insight_error_banner.tsx
@@ -20,20 +20,20 @@ export function AiInsightErrorBanner({ error, onRetry }: AiInsightErrorBannerPro
       announceOnMount
       color="danger"
       iconType="alert"
-      title={i18n.translate('observabilityAgentBuilder.aiInsight.errorTitle', {
+      title={i18n.translate('xpack.observability.aiInsight.errorTitle', {
         defaultMessage: 'Failed to generate AI insight',
       })}
       data-test-subj="AiInsightErrorBanner"
     >
       <p>
-        {i18n.translate('observabilityAgentBuilder.aiInsight.errorMessage', {
+        {i18n.translate('xpack.observability.aiInsight.errorMessage', {
           defaultMessage: 'The AI insight could not be generated: {error}',
           values: { error },
         })}
       </p>
       {onRetry && (
         <EuiButton size="s" onClick={onRetry} data-test-subj="AiInsightErrorBannerRetryButton">
-          {i18n.translate('observabilityAgentBuilder.aiInsight.retryButton', {
+          {i18n.translate('xpack.observability.aiInsight.retryButton', {
             defaultMessage: 'Try again',
           })}
         </EuiButton>
