+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/components/ai_insight/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export {
+  AiInsight,
+  type AiInsightProps,
+  type AiInsightResponse,
+  type AiInsightAttachment,
+} from './ai_insight';
+export { AiInsightErrorBanner, type AiInsightErrorBannerProps } from './ai_insight_error_banner';
+export { StartConversationButton } from './start_conversation_button';
+export { RobotIcon, robotIconType } from './robot_icon';
