+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/components/ai_insight/robot_icon.tsx
