--- x-pack/solutions/observability/plugins/observability_agent_builder/public/components/log_ai_insight/translations.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-import { i18n } from '@kbn/i18n';
-
-export const explainLogMessageButtonLabel = i18n.translate(
-  'xpack.observabilityAgentBuilder.aiInsight.explainLogMessageButtonLabel',
-  {
-    defaultMessage: 'Explain this log entry',
-  }
-);
-
-export const explainLogMessageDescription = i18n.translate(
-  'xpack.observabilityAgentBuilder.aiInsight.explainLogMessageDescription',
-  {
-    defaultMessage: 'Get helpful insights from our Elastic AI Agent',
-  }
-);
