+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/index.ts
@@ -22,5 +22,9 @@ export type {
   ObservabilityAgentBuilderPluginStartDependencies,
 } from './types';
 
+export type { AlertAiInsightProps, ErrorSampleAiInsightProps } from './components/insights';
+
+export { RobotIcon, robotIconType } from './components/ai_insight';
+
 export const plugin = (initializerContext: PluginInitializerContext) =>
   new ObservabilityAgentBuilderPlugin(initializerContext);
