+++ x-pack/solutions/observability/plugins/observability_agent_builder/tsconfig.json
@@ -35,7 +35,6 @@
     "@kbn/management-settings-ids",
     "@kbn/logging",
     "@kbn/ai-assistant-common",
-    "@kbn/ai-insights",
     "@kbn/kibana-react-plugin",
     "@kbn/shared-ux-utility",
     "@kbn/discover-shared-plugin",
