+++ x-pack/solutions/observability/plugins/serverless_observability/moon.yml
@@ -37,7 +37,7 @@ dependsOn:
   - '@kbn/deeplinks-management'
   - '@kbn/ai-assistant-common'
   - '@kbn/management-settings-ids'
-  - '@kbn/ai-insights'
+  - '@kbn/observability-agent-builder-plugin'
 tags:
   - plugin
   - prod
