+++ x-pack/solutions/observability/plugins/serverless_observability/public/navigation_tree.ts
@@ -27,7 +27,7 @@ const LazyIconProductCloudInfra = lazy(() =>
   }))
 );
 const LazyAgentBuilderIcon = lazy(() =>
-  import('@kbn/ai-insights/src/icons/robot_icon').then(({ RobotIcon }) => ({
+  import('@kbn/observability-agent-builder-plugin/public').then(({ RobotIcon }) => ({
     default: RobotIcon,
   }))
 );
