+++ x-pack/solutions/observability/plugins/serverless_observability/tsconfig.json
@@ -34,6 +34,6 @@
     "@kbn/deeplinks-management",
     "@kbn/ai-assistant-common",
     "@kbn/management-settings-ids",
-    "@kbn/ai-insights",
+    "@kbn/observability-agent-builder-plugin"
   ]
 }
